requirements_enhanced.txt
# Enhanced requirements with security and performance improvements

# Core dependencies
numpy>=1.21.0
PyQt5>=5.15.0
pyaudio>=0.2.11
sounddevice>=0.4.4
soundfile>=0.10.3
pygame>=2.1.0
requests>=2.26.0
psutil>=5.8.0

# Speech processing
openai-whisper>=1.0.0
pyttsx3>=2.90
gTTS>=2.2.3
elevenlabs>=0.2.0

# Browser automation
playwright>=1.30.0

# System utilities
pywin32>=228; platform_system == "Windows"
pycaw>=20181226; platform_system == "Windows"
wmi>=1.5.1; platform_system == "Windows"

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=4.0.0

# Visualization
PyQt5>=5.15.0
numpy>=1.21.0

# Security enhancements
cryptography>=3.4.8
pydantic>=1.10.0
python-jose>=3.3.0

# Configuration and validation
PyYAML>=6.0
marshmallow>=3.19.0

# Performance monitoring
prometheus-client>=0.15.0
memory-profiler>=0.60.0

# Database (optional)
sqlite3  # Built-in
sqlalchemy>=1.4.0

# Async support
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
pre-commit>=2.20.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Logging and monitoring
structlog>=22.0.0
sentry-sdk>=1.9.0